requirements.txt
# Core RAG Libraries

llama-index==0.10.53
openai==1.35.13
python-dotenv==1.0.1
chromadb==0.5.3

# Data Loaders & Integrations

beautifulsoup4==4.12.3
pypdf==4.2.0
llama-index-readers-web==0.1.12
llama-index-vector-stores-chroma==0.1.9

# Config & Scripts

PyYAML==6.0.1

# Evaluation

ragas==0.1.9

# Web Application

streamlit==1.37.0